Sampling

It is possible to sample trajectories of any fully defined StateSpaceModel. This can be done either in-place, or out-of-place:

Random.rand! — Function

Random.rand!([rng::AbstractRNG], out::AbstractMatrix, m::StateSpaceModel,
             x0::AbstractVector)
            
Random.rand!([rng::AbstractRNG], out_x::AbstractMatrix, out_y::AbstractMatrix,
             m::StateSpaceModel, x0::AbstractVector)

Sample a trajectory of the state space model m starting from x0, store the state trajectory in out_x (if out_x provided) and store the observations in out or out_y.

Random.rand!([rng::AbstractRNG], out::Abstract3Tensor, m::StateSpaceModel,
             x0::AbstractMatrix)

Random.rand!(rng::AbstractRNG, out_x::Abstract3Tensor, out_y::Abstract3Tensor,
             m::StateSpaceModel, x0::AbstractMatrix)

Sample a batch of trajectories of the state space model m, each starting from a separate x0[:, i], store the state trajectory in out_x (if out_x provided) and store the observations in out or out_y.

Note

x0 will be stored in out_x (if provided) as the first state and y0 (sampled from x0) will be stored in out_y (or out) as the first observation.

Base.rand — Function

Random.rand([rng::AbstractRNG], m::StateSpaceModel, x0::AbstractVector,
            num_obs::Int)

Sample a trajectory of the state space model m starting from x0 and return the state trajectory and the observations.

Random.rand([rng::AbstractRNG], m::StateSpaceModel, x0::AbstractMatrix,
            num_obs::Int)

Sample a batch of trajectories of the state space model m, each starting from a separate x0[:, i] and return the state trajectories and the observations.

Note

num_obs is the number of observations to be sampled, excluding the initial state. However, the initial state (as well as the "zeroth" observation) will be prepended to the output arrays.

The trajectories can be sampled either individually or in batches—the appropriate mode will be inferred from the dimension of the state space array x0. For in-place computations the number of timepoints (num_obs) will be inferred from the size of the output containers.

Out-of-place sampling will always return the sampled underlying state xx together with the observations yy.

For in-place sampling the underlying state can be discarded by passing only a container intended to store the observations yy.

Example

In the following example we first show how to sample a single trajectory and then a batch of trajectories.

using Distributions
using IMUDevStateSpaceModels
using LinearAlgebra

function constant_acceleration_model_2d(δt)
    # time-evolution of acceleration & velocity
    F = [1.0 δt;
         0.0 1.0]
    Q = [(δt^3)/3 (δt^2)/2;
         (δt^2)/2 δt]

    H = [1.0 0.0;]
    R = [1.0;;]
    return LinGsnSSM(Float32; F, Q, H, R)
end

m = constant_acceleration_model_2d(0.01)

╭──────────────────────────────────────────────────────────────────────────────╮
│  Linear Gaussian state space model                                           │
│  xₜ₊₁ = F xₜ + wₜ,  wₜ∼ N(0,Q),                                              │
│  yₜ = H xₜ + vₜ,  vₜ∼ N(0,R),                                                │
│  where xₜ is the unobserved, underlying state and yₜ are the observations.   │
│                                                                              │
╰──────────────────────────────────────────────────────────────────────────────╯

╭───────────────╮
│  State space  │
╰───────────────╯

State dimension: 2

    ╭──── Matrix ──────────╮         ╭──── Matrix ─────────────────────╮
    │                      │         │                                 │
F = │                      │     Q = │                                 │
    │         (1)   (2)    │         │             (1)         (2)     │
    │                      │         │                                 │
    │   (1)   1.0   0.01   │         │   (1)   3.3333333e-7   5.0e-5   │
    │   (2)   0.0   1.0    │         │   (2)   5.0e-5         0.01     │
    │                      │         │                                 │
    │                      │         │                                 │
    ╰─────────── 2 × 2  ───╯         ╰────────────────────── 2 × 2  ───╯

(where xₜ₊₁ =  F xₜ + wₜ,  wₜ∼ N(0,Q))                                          

╭─────────────────────╮
│  Observation space  │
╰─────────────────────╯

Observation dimension: 1

    ╭──── Matrix ─────────╮         ╭──── M... ─────╮
    │                     │         │               │
H = │                     │     R = │               │
    │         (1)   (2)   │         │         (1)   │
    │                     │         │               │
    │   (1)   1.0   0.0   │         │   (1)   1.0   │
    │                     │         │               │
    ╰────────── 1 × 2  ───╯         ╰────── 1... ───╯

(where yₜ =  H xₜ + vₜ,  vₜ∼ N(0,R))                                            

ℙx₀ = MvNormal(zeros(size(m, :state)), I)
num_timepoints = 100

# sample a single trajectory
xx_single, yy_single = rand(m, rand(ℙx₀), num_timepoints)

([-1.1577879331642373 -1.1687804387059646 … -2.9795883167638255 -3.0051932843926976; -1.1447818162175736 -1.0239228385566301 … -2.6210478484907727 -2.640609757280285], [-1.8563717049215795 -1.9431467992930513 … -2.6586860331425246 -3.877742124623832])

# sample a batch of trajectories
num_samples = 10
x0s = rand(ℙx₀, num_samples)
xx_batch, yy_batch = rand(m, x0s, num_timepoints)

([1.0799741658071393 0.856262666674467 … 0.5563585614867618 0.06429568386999036; -0.20700593285365299 0.35104566569822804 … 1.2881483918836902 -1.388233355715278;;; 1.0780119476307841 0.8596637232918115 … 0.5703492703243294 0.05021467300965722; -0.2383388934405759 0.21360635580520568 … 1.4787146658429502 -1.4587023995112356;;; 1.0750241400625589 0.8627402808999038 … 0.585093439619156 0.03651104046353095; -0.3358607790997256 0.4514682431190923 … 1.5579348927147179 -1.335583471888248;;; … ;;; 0.9890871405748226 1.0684557627564621 … 1.027264999348332 -1.7407479417554965; -0.9166261820674696 -0.5364955356910908 … 0.4481722462619675 -2.153620125366756;;; 0.9796246604978499 1.0625293177931654 … 1.0308271667838047 -1.7620705221818895; -0.9176978683006088 -0.645596533713281 … 0.36124014224152085 -2.1395334846541414;;; 0.9703008712132491 1.0567159987981545 … 1.0338229567300965 -1.783813409873531; -0.8765928811536683 -0.5149939798343562 … 0.2550710514290715 -2.1834121232298087], [2.371368562956037 2.24218851559568 … -0.3963028794512895 -1.8847429580643615;;; 0.5998724580727322 1.7859422368871172 … -0.047949480326304195 1.1628162667844242;;; 0.9699250429958667 1.0387131339887343 … 0.15029267232345145 -1.3847426983568683;;; … ;;; 1.047481930350811 0.6228198579802062 … 1.7964794991820854 -3.244969200266177;;; 0.3570141390230209 1.5906944337008508 … 0.01924272293538798 -0.9963486030152683;;; 0.8941676928385291 1.5320159840521936 … 0.27650721608460815 -0.2744157092860031])