State Space Models

Interface

All State Space Models defined in this package must adhere to the interface defined by the StateSpaceModel below.

IMUDevStateSpaceModels.StateSpaceModel — Type

Parent type to all State Space Models. T is the eltype of state.

In this package we deal with models of the form:

\[\begin{cases} x_{k+1} &= f(x_k) + ω_k,\qquad ω_k \sim Ω,\\ y_k &= g(x_k) + δ_k,\qquad δ_k \sim Δ, \end{cases}\]

where

$x_k$ is the state,

$y_k$ is the observation,

$f$ is the state transition function,

$g$ is the observation function,

$ω_k$ is the state noise,

$δ_k$ is the observation noise,

$Ω$ is the distribution of the state noise, and

$Δ$ is the distribution of the observation noise.

In particular, we do not deal with control inputs and we assume a linearly additive noise model.

Warning

We assume that the state is given by a vector (even a scalar-valued state is represented as a vector of length 1). The same applies to the observations. (At least for now... We might change this in the future, depending on the needs that will arise.)

The Julia type used for representation of the state and observation spaces should be known at compile time:

Base.eltype — Method

Base.eltype(::StateSpaceModel{T})

eltype of the state.

Note

Note however, that this does not necessarily prevent auto-promotion of variables. For instance, if state_transition below was called on a StateSpaceModel with Float32 eltype and a DualNumber Dual{Float32}, we would still expect each dimension of the state to be auto-promoted to Dual{Float32}.

Functions that must be implemented

Each concrete implementation of StateSpaceModel must implement the following functions defining its dynamics:

IMUDevStateSpaceModels.state_transition — Function

state_transition(x, m::StateSpaceModel)

State transition function $f$ of the StateSpaceModel m.

Important

Any concrete implementation of the StateSpaceModel must override this function.

IMUDevStateSpaceModels.observation_emission — Function

observation_emission(x, m::StateSpaceModel)

Observation function $g$ of the StateSpaceModel m.

Important

Any concrete implementation of the StateSpaceModel must override this function.

IMUDevStateSpaceModels.state_noise — Method

state_noise(m::StateSpaceModel)

State noise generator $Ω$ of the StateSpaceModel m.

Important

Any concrete implementation of the StateSpaceModel must override this function.

IMUDevStateSpaceModels.observation_noise — Method

observation_noise(m::StateSpaceModel)

Observation noise generator $Δ$ of the StateSpaceModel m.

Important

Any concrete implementation of the StateSpaceModel must override this function.

As well as functions defining the size of its state space and observation space.

Base.size — Method

Base.size(m::StateSpaceModel)

Return lengths of the state and observation vectors.

Base.size — Method

Base.size(m::StateSpaceModel, s::Union{Symbol,Val{:state},Val{:observation},Val{:obs})

Length of the state or observation vectors depending on the value of s:

s=:state or s=Val{:state}: length of the state,

s=:observation or s=:obs or s=Val{:observation} or s=Val{:obs}: length of the observation.

Important

Any concrete implementation of the StateSpaceModel must override functions Base.size(m::StateSpaceModel, ::Val{:state}) and Base.size(m::StateSpaceModel, ::Union{Val{:observation},Val{:obs}}).

Functions that will work out of the box

Based on these 6 definitions (4 dynamics definitions and 2 size definitions) the following evolution function will work out of the box:

IMUDevStateSpaceModels.step — Function

step(rng::AbstractRNG, m::StateSpaceModel,
     x::Union{<:AbstractVector,<:AbstractMatrix};
     state_noise_gen=state_noise(m), obs_noise_gen=observation_noise(m))

Perform a single (random) step:

\[\begin{cases} x° &= f(x_k) + ω_k,\qquad ω_k \sim Ω,\\ y° &= g(x°) + δ_k,\qquad δ_k \sim Δ, \end{cases}\]

for a single state $xₖ :=$ x if x is a vector, or

for a batch of states $xₖ⁽ⁱ⁾ :=$ x[:, i] if x is a matrix.

Under the hood it will call these two functions:

IMUDevStateSpaceModels.state_step — Function

state_step(rng::AbstractRNG, m::StateSpaceModel,
           x::Union{<:AbstractVector, <:AbstractMatrix};
           noise_gen=state_noise(m))

Perform a single (random) step:

\[f(x_k) + ω_k,\quad ω_k \sim Ω\]

for a single state $xₖ :=$ x if x is a vector, or

for a batch of states $xₖ⁽ⁱ⁾ :=$ x[:, i] if x is a matrix.

IMUDevStateSpaceModels.obs_step — Function

obs_step(rng::AbstractRNG, m::StateSpaceModel,
         x::Union{<:AbstractVector, <:AbstractMatrix};
         noise_gen=observation_noise(m))

Perform a single (random) observation step:

\[g(x_k) + δ_k,\quad δ_k \sim Δ\]

for a single state $xₖ :=$ x if x is a vector, or

for a batch of states $xₖ⁽ⁱ⁾ :=$ x[:, i] if x is a matrix.

Gaussian State Space Models

Abstract type that defines the type of state and observation noise to be Gaussian. Any child struct will inherit the definitions of $Ω:=N(0,Q)$ and $Δ:=N(0,R)$.

IMUDevStateSpaceModels.GaussianStateSpaceModel — Type

Parent type to Gaussian State Space Models of the form:

\[\begin{cases} x_{k+1} &= f(x_k) + ω_k,\qquad ω_k \sim N(0, Q),\\ y_k &= g(x_k) + δ_k,\qquad δ_k \sim N(0, R), \end{cases}\]

where $Q$ and $R$ are the state and observation noise covariance matrices, and the remaining parameters are as in the description of the StateSpaceModel type.

IMUDevStateSpaceModels.state_noise — Method

state_noise(m::GaussianStateSpaceModel)

Return $N(0, Q)$: a (Gaussian) distribution of the state noise of the GaussianStateSpaceModel m.

IMUDevStateSpaceModels.observation_noise — Method

observation_noise(m::GaussianStateSpaceModel)

Return $N(0, R)$: a (Gaussian) distribution of the observation noise of the GaussianStateSpaceModel m.

Any concrete implementation must define the following two functions:

IMUDevStateSpaceModels.state_noise_cov — Function

state_noise_cov(m::GaussianStateSpaceModel)

Return the state noise covariance matrix $Q$ of the GaussianStateSpaceModel m.

Important

Any concrete implementation of the GaussianStateSpaceModel must override this function.

IMUDevStateSpaceModels.observation_noise_cov — Function

observation_noise_cov(m::GaussianStateSpaceModel)

Return the observation noise covariance matrix $R$ of the GaussianStateSpaceModel m.

Important

Any concrete implementation of the GaussianStateSpaceModel must override this function.

Note

Additionally, functions state_transition, observation_emission, and two versions of Base.size must be implemented to satisfy the requirements of the StateSpaceModel interface.

Linear Gaussian State Space Model

LinearGaussianStateSpaceModel is a concrete implementation of the GaussianStateSpaceModel (thus in particular, also of StateSpaceModel's).

IMUDevStateSpaceModels.LinearGaussianStateSpaceModel — Type

LinearGaussianStateSpaceModel([T::DataType]; F, Q, H, R)

State space model with linear dynamics, additive Gaussian noise and linear, Gaussian observations. It is given by the following set of equations:

\[\begin{cases} x_{t+1} &= F \times x_t + w_t,\qquad w_t \sim N(0, Q)\\ y_t &= H \times x_t + v_t,\qquad v_t \sim N(0, R) \end{cases}\]

Tip

You can use LinGsnSSM alias for LinearGaussianStateSpaceModel.

The following additional functions are defined for LinearGaussianStateSpaceModel:

IMUDevStateSpaceModels.transition_matrix — Function

transition_matrix(m::LinGsnSSM)

Return state evolution matrix $F$ of the LinearGaussianStateSpaceModel m.

IMUDevStateSpaceModels.observation_matrix — Function

observation_matrix(m::LinGsnSSM)

Return the observation matrix $H$ of the LinearGaussianStateSpaceModel m.

Example of a Non-linear Non-Gaussian State Space Model

More general models must be implemented by the user by adhering to the StateSpaceModel interface. Below is an example of a 1-dimensional state evolving as a double-well potential with Cauchy noise and non-linear observation model perturbed by Poisson noise:

using Distributions
using IMUDevStateSpaceModels
using Plots

const SSMs = IMUDevStateSpaceModels

"""
    MyNonLinearStateSpaceModel()

State space model with double-well type transition function, non-linear
observation operator and non-Gaussian noise, given by:

    x_{k+1} = x_k - θ(x_k)^2((x_k)^2 - μ) + σ_k,    σ_k ∼ Cauchy(x₀, γ),
    y_k     = ceil(x_k)              + ν_k,    ν_k ∼ Poisson(λ)
"""
struct MyNonLinearStateSpaceModel <: StateSpaceModel{Float64}
    θ::Float64
    μ::Float64
    x₀::Float64
    γ::Float64
    λ::Int64
end

const MySSM = MyNonLinearStateSpaceModel

Base.size(::MySSM, ::Union{Val{:state}}) = 1
Base.size(::MySSM, ::Union{Val{:observation},Val{:obs}}) = 1
SSMs.state_noise(m::MySSM) = Product([Cauchy(m.x₀, m.γ)])
SSMs.observation_noise(m::MySSM) = Product([Poisson(m.λ)])

SSMs.state_transition(x, m::MySSM) = x - m.θ .* x .* (x .^ 2 .- m.μ)
SSMs.observation_emission(x, ::MySSM) = ceil.(x)

Tip

Notice that despite the model being 1-dimensional we don't treat it as a scalar-valued process, but instead, as a 1-element vector-valued process. In particular, we use the Distributions.jl Product to compose Univariate distributions into a Multivariate one.

Once defined, we can define a model and, for instance, sample and plot trajectories (see Sampling and Plotting sections for more details):

using Plots

model = MySSM(0.01, 1.0, 0.0, 0.001, 1)

x₀ = [1.0]
num_timepoints = 100
xx, yy = rand(model, x₀, num_timepoints)

plot(trajectoryplot(0:num_timepoints, yy; seriestype=:scatter, label="", title="Observations"),
     trajectoryplot(0:num_timepoints, xx; seriestype=:scatter, label="", title="Underlying state"))